Fix dynamic-level-bytes heuristic
The implementation of dynamic-level-bytes was incomplete: we were
dynamically determining the base level and the size for each level, but
the base level wasn't being utilized to compute the compaction output
level when compacting from L0. The end result is that we'd determine the
LSM only needed 2 levels, but then compact from L0 -> L1, thus
invalidating the dyanmic-level-bytes heuristic (which stops at the first
non-empty level).

Also fixed the target file size calculation to account for the dynamic
base level.

Fixes #108